fix x64 warning
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@5161 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ applications/mp42ts/main.c
@@ -368,7 +368,7 @@ static GF_Err mp4_input_ctrl(GF_ESInterface *ifce, u32 act_type, void *param)
 			cue = (GF_WebVTTCue *)gf_list_get(cues, 0);
 			if (cue) {
 				pck.data = cue->text;
-				pck.data_len = strlen(cue->text)+1;
+				pck.data_len = (u32)strlen(cue->text)+1;
 			} else {
 				pck.data = NULL;
 				pck.data_len = 0;
